rename breath to breathe, fixing a copied typo
